+++ Source/DotNetWorkQueue.Transport.LiteDB/Basic/LiteDbMessageQueueInit.cs
@@ -114,7 +114,6 @@ namespace DotNetWorkQueue.Transport.LiteDb.Basic
             container.Register<MessageDeQueue>(LifeStyles.Singleton);
             container.Register<IReceiveMessages, LiteDbQueueReceiveMessages>(LifeStyles.Transient);
             container.Register<ReceiveMessage>(LifeStyles.Singleton);
-            container.Register<RollbackMessage>(LifeStyles.Singleton);
 
             container.Register<IResetHeartBeat, ResetHeartBeat<int>>(LifeStyles.Singleton);
             container.Register<ISendHeartBeat, SendHeartBeat<int>>(LifeStyles.Singleton);